Started refactoring spells
